+++ packages/fluentui/react-northstar/src/components/Avatar/Avatar.tsx
@@ -82,7 +82,7 @@ export const Avatar = (React.forwardRef<HTMLDivElement, AvatarProps>((props, ref
     debugName: Avatar.displayName,
     rtl: context.rtl,
   });
-  const { classes, styles: resolvedStyles } = useStyles(Avatar.displayName, {
+  const { classes } = useStyles(Avatar.displayName, {
     className: avatarClassName,
     mapPropsToStyles: () => ({ size, square }),
     mapPropsToInlineStyles: () => ({
@@ -103,8 +103,6 @@ export const Avatar = (React.forwardRef<HTMLDivElement, AvatarProps>((props, ref
         avatar: !square,
         title: name,
         size,
-        // remove in upcoming breaking change
-        styles: resolvedStyles.image,
       }),
   });
 
@@ -112,7 +110,6 @@ export const Avatar = (React.forwardRef<HTMLDivElement, AvatarProps>((props, ref
     defaultProps: () =>
       getA11Props('icon', {
         title: name,
-        styles: resolvedStyles.icon,
         size,
         square,
       }),
@@ -126,7 +123,6 @@ export const Avatar = (React.forwardRef<HTMLDivElement, AvatarProps>((props, ref
         title: name,
         size,
         square,
-        styles: resolvedStyles.label,
       }),
   });
 
@@ -140,8 +136,6 @@ export const Avatar = (React.forwardRef<HTMLDivElement, AvatarProps>((props, ref
         defaultProps: () =>
           getA11Props('status', {
             size,
-            // remove in upcoming breaking change
-            styles: resolvedStyles.status,
           }),
         overrideProps: (predefinedProps: AvatarStatusProps) => ({
           variables: mergeVariablesOverrides(variables, predefinedProps.variables),