+++ packages/react-components/react-button/src/components/MenuButton/MenuButton.test.tsx
@@ -10,6 +10,7 @@ describe('MenuButton', () => {
   isConformant({
     Component: MenuButton as React.FunctionComponent<MenuButtonProps>,
     displayName: 'MenuButton',
+    disabledTests: ['component-has-static-classname-exported'],
     testOptions: {
       'has-static-classnames': [
         {