+++ packages/react-components/react-checkbox/src/components/Checkbox/Checkbox.test.tsx
@@ -15,6 +15,7 @@ describe('Checkbox', () => {
   isConformant({
     Component: Checkbox,
     displayName: 'Checkbox',
+    disabledTests: ['component-has-static-classname-exported'],
     primarySlot: 'input',
     testOptions: {
       'has-static-classnames': [