+++ packages/react-components/react-radio/src/stories/RadioGroupHorizontal.stories.tsx
@@ -1,12 +1,13 @@
 import * as React from 'react';
 import { Label } from '@fluentui/react-label';
+import { tokens } from '@fluentui/react-theme';
 import { useId } from '@fluentui/react-utilities';
 import { Radio, RadioGroup } from '../index';
 
 export const Horizontal = () => {
   const labelId = useId('label');
   return (
-    <div style={{ display: 'flex', flexDirection: 'column', gap: '8px' }}>
+    <div style={{ display: 'grid', gridRowGap: tokens.spacingVerticalS }}>
       <Label id={labelId}>Favorite Fruit</Label>
       <RadioGroup layout="horizontal" aria-labelledby={labelId}>
         <Radio value="apple" label="Apple" />