+++ packages/react-components/react-spinbutton/src/components/SpinButton/SpinButton.types.ts
@@ -32,20 +32,25 @@ export type SpinButtonSlots = {
 /**
  * SpinButton Props
  */
-export type SpinButtonProps = Omit<ComponentProps<Partial<SpinButtonSlots>, 'input'>, 'onChange' | 'size'> & {
+export type SpinButtonProps = Omit<
+  ComponentProps<Partial<SpinButtonSlots>, 'input'>,
+  'defaultValue' | 'onChange' | 'size' | 'value'
+> & {
   /**
    * Controls the colors and borders of the input.
    * @default 'outline'
    */
-  appearance?: 'outline' | 'underline' | 'filledDarker' | 'filledLighter';
+  appearance?: 'outline' | 'underline' | 'filled-darker' | 'filled-lighter';
 
   /**
    * Initial value of the control (assumed to be valid). Updates to this prop will not be respected.
    *
    * Use this if you intend for the SpinButton to be an uncontrolled component which maintains its
    * own value. For a controlled component, use `value` instead. (Mutually exclusive with `value`.)
+   *
+   * Use `null` to indicate the control has no value.
    */
-  defaultValue?: number;
+  defaultValue?: number | null;
 
   /**
    * String representation of `value`.
@@ -108,18 +113,17 @@ export type SpinButtonProps = Omit<ComponentProps<Partial<SpinButtonSlots>, 'inp
   stepPage?: number;
 
   /**
-   * Strings for localizing text in the control.
-   */
-  strings?: SpinButtonStrings;
-
-  /**
    * Current value of the control (assumed to be valid).
    *
    * Only provide this if the SpinButton is a controlled component where you are maintaining its
    * current state and passing updates based on change events; otherwise, use the `defaultValue`
-   * property. (Mutually exclusive with `defaultValue`.)
+   * property.
+   *
+   * Use `null` to indicate the control has no value.
+   *
+   * Mutually exclusive with `defaultValue`.
    */
-  value?: number;
+  value?: number | null;
 };
 
 /**
@@ -147,23 +151,9 @@ export type SpinButtonChangeEvent =
   | React.KeyboardEvent<HTMLInputElement>;
 
 export type SpinButtonOnChangeData = {
-  value?: number;
+  value?: number | null;
   displayValue?: string;
 };
 
 export type SpinButtonSpinState = 'rest' | 'up' | 'down';
 export type SpinButtonBounds = 'none' | 'min' | 'max' | 'both';
-
-export type SpinButtonStrings = {
-  /**
-   * Label applied to the increment button.
-   * Can include the token "\{step\}" which will be replaced with the value of the `step` prop.
-   */
-  incrementButtonLabel: string;
-
-  /**
-   * Label applied to the decrement button.
-   * Can include the token "\{step\}" which will be replaced with the value of the `step` prop.
-   */
-  decrementButtonLabel: string;
-};