+++ packages/react-components/react-spinbutton/src/utils/clamp.ts
@@ -1,12 +1,6 @@
-export const clampWhenInRange = (oldValue: number, newValue: number, min?: number, max?: number): number => {
-  // When oldValue is in the range of [min, max] clamp newValue.
-  // Don't clamp values outside this range so users get a
-  // more natural behavior. For example, if the range is [5, 15]
-  // and the user types 1 into the input we don't want to clamp
-  // the value when they next press the increment button because
-  // clamping would snap the value to 5 rather than increment to 2.
-  let nextValue = newValue;
-  if (min !== undefined && oldValue >= min) {
+export const clamp = (value: number, min?: number, max?: number): number => {
+  let nextValue = value;
+  if (min !== undefined) {
     if (max !== undefined && min > max) {
       const error = new Error();
       if (process.env.NODE_ENV !== 'production') {
@@ -15,18 +9,18 @@ export const clampWhenInRange = (oldValue: number, newValue: number, min?: numbe
           [
             `"min" value "${min}" is greater than "max" value "${max}".`,
             '"min" must be less than or equal to "max".',
-            `Returning value "${newValue}".`,
+            `Returning value "${value}".`,
             error.stack,
           ].join(),
         );
       }
-      return newValue;
+      return value;
     }
 
     nextValue = Math.max(min, nextValue);
   }
 
-  if (max !== undefined && oldValue <= max) {
+  if (max !== undefined) {
     nextValue = Math.min(max, nextValue);
   }