+++ packages/react-components/react-text/src/Caption1.ts
@@ -1,0 +1,1 @@
+export * from './components/Caption1/index';